Migration to Swift 4 incl. CocoaPods dependency management
