+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/.gitignore
@@ -1,0 +1,1 @@
+Pods
