+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/Podfile
@@ -1,0 +1,13 @@
+# Uncomment this line to define a global platform for your project
+# platform :ios, '10.0'
+
+target 'SAPJamSDK' do
+  # Comment this line if you're not using Swift and don't want to use dynamic frameworks
+  use_frameworks!
+
+  # Pods for SAPJamSDK
+  
+  pod 'KeychainAccess'
+  pod 'OAuthSwift'
+  
+end
