+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDK.xcodeproj/xcuserdata/i850769.xcuserdatad/xcschemes/xcschememanagement.plist
