+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDK.xcworkspace/xcshareddata/SAPJamSDK.xcscmblueprint
