+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDK.xcworkspace/xcuserdata/i850769.xcuserdatad/UserInterfaceState.xcuserstate
