+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDK/Assets.xcassets/first.imageset/Contents.json
