+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDK/Assets.xcassets/second.imageset/Contents.json
