+++ iOS_Integrations/mobile-sdk-ios/Swift 4.x/SAPJamSDK/SAPJamSDKUITests/SAPJamSDKUITests.swift
